feat: added basic support for integration tests
With this there can now be test written that call against the endpoints of the application.
