+++ package.json
@@ -36,8 +36,9 @@
     "migrate": "node node_modules/migrate/bin/migrate",
     "prestart": "npm run -s build",
     "test": "eslint src",
-    "test:unit": "jest -c test/unit/jest.conf.js",
-    "setup": "npm run restore; npm run migrate",
+    "test:unit": "NODE_ENV=test jest -c test/unit/jest.conf.js",
+    "test:unit:watch": "NODE_ENV=test jest -c test/unit/jest.conf.js --watch",
+    "setup": "npm run restore7; npm run migrate",
     "lint": "eslint --ext .js,.ts packages src migrations scripts --fix",
     "lint-gql": "graphql-schema-linter {src,packages}/**/*.graphqls"
   },
@@ -55,10 +56,10 @@
     "@2fd/graphdoc": "^2.4.0",
     "@elastic/elasticsearch": "^7.3.0",
     "@google-cloud/storage": "^3.0.3",
-    "@storefront-api/lib": "1.0.0-rc2",
     "@storefront-api/default-catalog": "1.0.0-rc2",
     "@storefront-api/default-img": "1.0.0-rc2",
     "@storefront-api/default-vsf": "1.0.0-rc2",
+    "@storefront-api/lib": "1.0.0-rc2",
     "@storefront-api/platform": "1.0.0-rc2",
     "@storefront-api/platform-abstract": "1.0.0-rc2",
     "@storefront-api/platform-magento1": "1.0.0-rc2",
@@ -108,6 +109,7 @@
     "semver": "^6.3.0",
     "sharp": "^0.23.4",
     "soap": "^0.25.0",
+    "supertest": "^4.0.2",
     "syswide-cas": "latest",
     "ts-node": "^8.1.0",
     "tsconfig-paths": "^3.9.0",
@@ -127,6 +129,7 @@
     "@types/node": "^11.13.4",
     "@types/redis": "^2.8.14",
     "@types/semver": "^6.2.0",
+    "@types/supertest": "^2.0.8",
     "@typescript-eslint/eslint-plugin": "^1.13.0",
     "@typescript-eslint/parser": "^1.13.0",
     "apollo-server-express": "^2.9.12",
