+++ test/unit/jest.conf.js
@@ -7,15 +7,16 @@ module.exports = {
   ],
   testMatch: [
     '<rootDir>/src/**/test/unit/**/*.spec.(js|ts)',
+    '<rootDir>/packages/**/test/unit/**/*.spec.(js|ts)'
   ],
   transform: {
     '^.+\\.js$': '<rootDir>/node_modules/ts-jest',
-    '^.+\\.ts$': '<rootDir>/node_modules/ts-jest',
+    '^.+\\.ts$': '<rootDir>/node_modules/ts-jest'
   },
   coverageDirectory: '<rootDir>/test/unit/coverage',
   collectCoverageFrom: [
     'src/**/*.{js,ts}',
-    '!src/**/types/*.{js,ts}',
+    '!src/**/types/*.{js,ts}'
   ],
   moduleNameMapper: {
     '^src(.*)$': '<rootDir>/src$1'
